refactor:♻️FieldBatchInput型をdomain/dtoに移動(#4)
Infrastructure→Application参照違反を解消するため、FieldBatchInput型を
application/usecaseからdomain/dtoに移動。Domain層同士の参照は許容される。
