+++ internal/features/field/infrastructure/repository/field_integration_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/google/uuid"
 	"github.com/jackc/pgx/v5/pgxpool"
 	"github.com/mktkhr/field-manager-api/internal/features/field/domain/entity"
-	importusecase "github.com/mktkhr/field-manager-api/internal/features/import/application/usecase"
+	importdto "github.com/mktkhr/field-manager-api/internal/features/import/domain/dto"
 )
 
 var testDB *pgxpool.Pool
@@ -140,7 +140,7 @@ func TestFieldRepository_UpsertBatch_EmptyFeatures_Integration(t *testing.T) {
 	repo := NewFieldRepository(testDB)
 
 	// 空の入力リストでUpsertBatch
-	err := repo.UpsertBatch(ctx, []importusecase.FieldBatchInput{})
+	err := repo.UpsertBatch(ctx, []importdto.FieldBatchInput{})
 	if err != nil {
 		t.Errorf("UpsertBatch() with empty features error = %v", err)
 	}
@@ -154,10 +154,10 @@ func TestFieldRepository_UpsertBatch_SingleFeature_Integration(t *testing.T) {
 	repo := NewFieldRepository(testDB)
 
 	fieldID := uuid.New()
-	input := importusecase.FieldBatchInput{
+	input := importdto.FieldBatchInput{
 		ID:       fieldID.String(),
 		CityCode: "163210",
-		Geometry: importusecase.FieldBatchGeometry{
+		Geometry: importdto.FieldBatchGeometry{
 			Type: "Polygon",
 			Coordinates: [][][]float64{
 				{
@@ -171,7 +171,7 @@ func TestFieldRepository_UpsertBatch_SingleFeature_Integration(t *testing.T) {
 		},
 	}
 
-	err := repo.UpsertBatch(ctx, []importusecase.FieldBatchInput{input})
+	err := repo.UpsertBatch(ctx, []importdto.FieldBatchInput{input})
 	if err != nil {
 		t.Errorf("UpsertBatch() error = %v", err)
 	}
@@ -194,10 +194,10 @@ func TestFieldRepository_UpsertBatch_WithSoilType_Integration(t *testing.T) {
 	repo := NewFieldRepository(testDB)
 
 	fieldID := uuid.New()
-	input := importusecase.FieldBatchInput{
+	input := importdto.FieldBatchInput{
 		ID:       fieldID.String(),
 		CityCode: "163210",
-		Geometry: importusecase.FieldBatchGeometry{
+		Geometry: importdto.FieldBatchGeometry{
 			Type: "Polygon",
 			Coordinates: [][][]float64{
 				{
@@ -209,7 +209,7 @@ func TestFieldRepository_UpsertBatch_WithSoilType_Integration(t *testing.T) {
 				},
 			},
 		},
-		SoilType: &importusecase.FieldBatchSoilType{
+		SoilType: &importdto.FieldBatchSoilType{
 			LargeCode:  "A",
 			MiddleCode: "A1",
 			SmallCode:  "A1a",
@@ -217,7 +217,7 @@ func TestFieldRepository_UpsertBatch_WithSoilType_Integration(t *testing.T) {
 		},
 	}
 
-	err := repo.UpsertBatch(ctx, []importusecase.FieldBatchInput{input})
+	err := repo.UpsertBatch(ctx, []importdto.FieldBatchInput{input})
 	if err != nil {
 		t.Errorf("UpsertBatch() with soil type error = %v", err)
 	}
@@ -240,10 +240,10 @@ func TestFieldRepository_UpsertBatch_WithPinInfo_Integration(t *testing.T) {
 
 	fieldID := uuid.New()
 	descriptiveStudyDataRaw := "2024-01-15"
-	input := importusecase.FieldBatchInput{
+	input := importdto.FieldBatchInput{
 		ID:       fieldID.String(),
 		CityCode: "163210",
-		Geometry: importusecase.FieldBatchGeometry{
+		Geometry: importdto.FieldBatchGeometry{
 			Type: "Polygon",
 			Coordinates: [][][]float64{
 				{
@@ -255,7 +255,7 @@ func TestFieldRepository_UpsertBatch_WithPinInfo_Integration(t *testing.T) {
 				},
 			},
 		},
-		PinInfoList: []importusecase.FieldBatchPinInfo{
+		PinInfoList: []importdto.FieldBatchPinInfo{
 			{
 				FarmerNumber:            "F001",
 				Address:                 "東京都渋谷区1-1-1",
@@ -269,7 +269,7 @@ func TestFieldRepository_UpsertBatch_WithPinInfo_Integration(t *testing.T) {
 		},
 	}
 
-	err := repo.UpsertBatch(ctx, []importusecase.FieldBatchInput{input})
+	err := repo.UpsertBatch(ctx, []importdto.FieldBatchInput{input})
 	if err != nil {
 		t.Errorf("UpsertBatch() with PinInfo error = %v", err)
 	}
@@ -290,10 +290,10 @@ func TestFieldRepository_UpsertBatch_InvalidFieldID_Integration(t *testing.T) {
 
 	repo := NewFieldRepository(testDB)
 
-	input := importusecase.FieldBatchInput{
+	input := importdto.FieldBatchInput{
 		ID:       "invalid-uuid",
 		CityCode: "163210",
-		Geometry: importusecase.FieldBatchGeometry{
+		Geometry: importdto.FieldBatchGeometry{
 			Type: "Polygon",
 			Coordinates: [][][]float64{
 				{
@@ -307,7 +307,7 @@ func TestFieldRepository_UpsertBatch_InvalidFieldID_Integration(t *testing.T) {
 		},
 	}
 
-	err := repo.UpsertBatch(ctx, []importusecase.FieldBatchInput{input})
+	err := repo.UpsertBatch(ctx, []importdto.FieldBatchInput{input})
 	if err == nil {
 		t.Error("UpsertBatch() with invalid field ID should return error")
 	}
@@ -321,10 +321,10 @@ func TestFieldRepository_UpsertBatch_InvalidGeometry_Integration(t *testing.T) {
 	repo := NewFieldRepository(testDB)
 
 	fieldID := uuid.New()
-	input := importusecase.FieldBatchInput{
+	input := importdto.FieldBatchInput{
 		ID:       fieldID.String(),
 		CityCode: "163210",
-		Geometry: importusecase.FieldBatchGeometry{
+		Geometry: importdto.FieldBatchGeometry{
 			Type: "Polygon",
 			Coordinates: [][][]float64{
 				{
@@ -336,7 +336,7 @@ func TestFieldRepository_UpsertBatch_InvalidGeometry_Integration(t *testing.T) {
 		},
 	}
 
-	err := repo.UpsertBatch(ctx, []importusecase.FieldBatchInput{input})
+	err := repo.UpsertBatch(ctx, []importdto.FieldBatchInput{input})
 	if err == nil {
 		t.Error("UpsertBatch() with invalid geometry should return error")
 	}
@@ -352,11 +352,11 @@ func TestFieldRepository_UpsertBatch_MultipleFeatures_Integration(t *testing.T) 
 	fieldID1 := uuid.New()
 	fieldID2 := uuid.New()
 
-	inputs := []importusecase.FieldBatchInput{
+	inputs := []importdto.FieldBatchInput{
 		{
 			ID:       fieldID1.String(),
 			CityCode: "163210",
-			Geometry: importusecase.FieldBatchGeometry{
+			Geometry: importdto.FieldBatchGeometry{
 				Type: "Polygon",
 				Coordinates: [][][]float64{
 					{
@@ -372,7 +372,7 @@ func TestFieldRepository_UpsertBatch_MultipleFeatures_Integration(t *testing.T) 
 		{
 			ID:       fieldID2.String(),
 			CityCode: "131016",
-			Geometry: importusecase.FieldBatchGeometry{
+			Geometry: importdto.FieldBatchGeometry{
 				Type: "Polygon",
 				Coordinates: [][][]float64{
 					{
@@ -418,10 +418,10 @@ func TestFieldRepository_UpsertBatch_UpdateExisting_Integration(t *testing.T) {
 	repo := NewFieldRepository(testDB)
 
 	fieldID := uuid.New()
-	input := importusecase.FieldBatchInput{
+	input := importdto.FieldBatchInput{
 		ID:       fieldID.String(),
 		CityCode: "163210",
-		Geometry: importusecase.FieldBatchGeometry{
+		Geometry: importdto.FieldBatchGeometry{
 			Type: "Polygon",
 			Coordinates: [][][]float64{
 				{
@@ -436,13 +436,13 @@ func TestFieldRepository_UpsertBatch_UpdateExisting_Integration(t *testing.T) {
 	}
 
 	// 最初のUpsert
-	err := repo.UpsertBatch(ctx, []importusecase.FieldBatchInput{input})
+	err := repo.UpsertBatch(ctx, []importdto.FieldBatchInput{input})
 	if err != nil {
 		t.Fatalf("First UpsertBatch() error = %v", err)
 	}
 
 	// 2回目のUpsert(同じIDで更新)
-	err = repo.UpsertBatch(ctx, []importusecase.FieldBatchInput{input})
+	err = repo.UpsertBatch(ctx, []importdto.FieldBatchInput{input})
 	if err != nil {
 		t.Errorf("Second UpsertBatch() error = %v", err)
 	}
@@ -487,16 +487,16 @@ func TestFieldRepository_UpsertBatch_TransactionError_Integration(t *testing.T) 
 
 	repo := NewFieldRepository(testDB)
 
-	input := importusecase.FieldBatchInput{
+	input := importdto.FieldBatchInput{
 		ID:       uuid.New().String(),
 		CityCode: "163210",
-		Geometry: importusecase.FieldBatchGeometry{
+		Geometry: importdto.FieldBatchGeometry{
 			Type:        "Polygon",
 			Coordinates: [][][]float64{{{139.6917, 35.6895}, {139.6920, 35.6895}, {139.6920, 35.6898}, {139.6917, 35.6898}, {139.6917, 35.6895}}},
 		},
 	}
 
-	err := repo.UpsertBatch(ctx, []importusecase.FieldBatchInput{input})
+	err := repo.UpsertBatch(ctx, []importdto.FieldBatchInput{input})
 	if err == nil {
 		t.Error("UpsertBatch() with cancelled context should return error")
 	}
@@ -510,16 +510,16 @@ func TestFieldRepository_UpsertBatch_EmptyGeometry_Integration(t *testing.T) {
 	repo := NewFieldRepository(testDB)
 
 	fieldID := uuid.New()
-	input := importusecase.FieldBatchInput{
+	input := importdto.FieldBatchInput{
 		ID:       fieldID.String(),
 		CityCode: "163210",
-		Geometry: importusecase.FieldBatchGeometry{
+		Geometry: importdto.FieldBatchGeometry{
 			Type:        "Polygon",
 			Coordinates: [][][]float64{}, // 空のコーディネート
 		},
 	}
 
-	err := repo.UpsertBatch(ctx, []importusecase.FieldBatchInput{input})
+	err := repo.UpsertBatch(ctx, []importdto.FieldBatchInput{input})
 	if err == nil {
 		t.Error("UpsertBatch() with empty geometry should return error")
 	}
@@ -534,10 +534,10 @@ func TestFieldRepository_UpsertBatch_TwoPointsGeometry_Integration(t *testing.T)
 
 	fieldID := uuid.New()
 	// 2点だけのポリゴン - geom.SetCoordsでエラーになるはず
-	input := importusecase.FieldBatchInput{
+	input := importdto.FieldBatchInput{
 		ID:       fieldID.String(),
 		CityCode: "163210",
-		Geometry: importusecase.FieldBatchGeometry{
+		Geometry: importdto.FieldBatchGeometry{
 			Type: "Polygon",
 			Coordinates: [][][]float64{
 				{
@@ -548,7 +548,7 @@ func TestFieldRepository_UpsertBatch_TwoPointsGeometry_Integration(t *testing.T)
 		},
 	}
 
-	err := repo.UpsertBatch(ctx, []importusecase.FieldBatchInput{input})
+	err := repo.UpsertBatch(ctx, []importdto.FieldBatchInput{input})
 	if err == nil {
 		t.Error("UpsertBatch() with two points polygon should return error")
 	}
