+++ internal/features/import/domain/dto/field_batch_input.go
@@ -1,9 +1,9 @@
-package usecase
+package dto
 
 import "time"
 
 // FieldBatchInput はバッチUPSERT用の入力データ
-// Consumer側(import機能)で型を定義し、Provider側(field機能)が実装する
+// Consumer側(import機能)のDomain層で型を定義し、Provider側(field機能)が実装する
 type FieldBatchInput struct {
 	ID          string
 	CityCode    string
